♻️ refactor: exams/serializes/error.py --> core/serializers/error.py
