+++ apps/exams/views/admin/deployments_list.py
@@ -21,7 +21,7 @@ from apps.exams.error_map import raise_error
 from apps.exams.serializers.admin.deployments_list import (
     AdminExamDeploymentListItemSerializer,
 )
-from apps.exams.serializers.error_serializers import ErrorResponseSerializer
+from apps.core.serializers.error import ErrorResponseSerializer
 from apps.exams.services.admin.deployments_list import (
     AdminDeploymentListService,
     InvalidAdminDeploymentListParams,
