+++ apps/exams/views/admin/deployments_router.py
@@ -17,7 +17,7 @@ from apps.exams.serializers.admin.deployments_create import (
     AdminExamDeploymentCreateRequestSerializer,
     AdminExamDeploymentCreateResponseSerializer,
 )
-from apps.exams.serializers.error_serializers import ErrorResponseSerializer
+from apps.core.serializers.error import ErrorResponseSerializer
 from apps.exams.views.admin.deployments_create import (
     AdminExamDeploymentCreateAPIView,
 )
