+++ apps/exams/views/admin/submissions_detail.py
@@ -10,14 +10,13 @@ from rest_framework.views import APIView
 
 from apps.core.utils.permissions import IsStaffRole
 from apps.exams.constants import ErrorMessages
-from apps.exams.error_map import raise_error
 from apps.exams.serializers.admin.submissions_delete import (
     AdminExamSubmissionDeleteResponseSerializer,
 )
 from apps.exams.serializers.admin.submissions_detail import (
     AdminExamSubmissionDetailResponseSerializer,
 )
-from apps.exams.serializers.error_serializers import ErrorResponseSerializer
+from apps.core.serializers.error import ErrorResponseSerializer
 from apps.exams.services.admin.submissions_delete import delete_exam_submission
 from apps.exams.services.admin.submissions_detail import get_admin_submission_detail
 from apps.exams.validators import parse_positive_int
