+++ apps/exams/views/student/deployments_take.py
@@ -11,7 +11,7 @@ from rest_framework.views import APIView
 
 from apps.exams.constants import ErrorMessages
 from apps.exams.serializers import TakeExamResponseSerializer
-from apps.exams.serializers.error_serializers import ErrorResponseSerializer
+from apps.core.serializers.error import ErrorResponseSerializer
 from apps.exams.services import build_take_exam_response, take_exam
 from apps.exams.validators import parse_positive_int
 from apps.exams.views.mixins import ExamsExceptionMixin
