+++ apps/exams/views/student/submissions_result.py
@@ -6,7 +6,7 @@ from rest_framework.permissions import IsAuthenticated
 
 from apps.exams.constants import ErrorMessages
 from apps.exams.models import ExamSubmission
-from apps.exams.serializers.error_serializers import ErrorResponseSerializer
+from apps.core.serializers.error import ErrorResponseSerializer
 from apps.exams.serializers.student.submissions_result import ExamSubmissionSerializer
 from apps.exams.services.student.submissions_result import get_exam_submission_detail
 from apps.exams.validators import parse_positive_int
